pyproject.toml
[project]
name = "agentdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=12.19.0",
    "azure-core>=1.36.0",
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "ddgs>=9.9.0",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.0.5",
    "langchain-azure-postgresql>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.2",
    "openinference-instrumentation-openai>=0.1.40",
    "psycopg>=3.2.12",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.12",
]
